requirements.txt
langchain
openai
chromadb
pypdf
pdfplumber
fastapi
uvicorn
python-dotenv
tiktoken